Move V2 out of prototype folder
