+++ app/plugins/router.js
@@ -63,16 +63,16 @@ const routes = [].concat(
   require('../routes/land-management/improved-grassland-soils/high/stocking-density'),
   require('../routes/land-management/improved-grassland-soils/high/soil-management'),
   require('../routes/land-management/improved-grassland-soils/high/high-end'),
-  require('../routes/prototypes/v2/search-crn'),
-  require('../routes/prototypes/v2/select-sbi'),
-  require('../routes/prototypes/v2/organisation-details'),
-  require('../routes/prototypes/v2/eligibility'),
-  require('../routes/prototypes/v2/standards'),
-  require('../routes/prototypes/v2/add-standards-parcels'),
-  require('../routes/prototypes/v2/choose-level'),
-  require('../routes/prototypes/v2/summary'),
-  require('../routes/prototypes/v2/submit'),
-  require('../routes/prototypes/v2/confirmation'),
+  require('../routes/v2/search-crn'),
+  require('../routes/v2/select-sbi'),
+  require('../routes/v2/organisation-details'),
+  require('../routes/v2/eligibility'),
+  require('../routes/v2/standards'),
+  require('../routes/v2/add-standards-parcels'),
+  require('../routes/v2/choose-level'),
+  require('../routes/v2/summary'),
+  require('../routes/v2/submit'),
+  require('../routes/v2/confirmation'),
   require('../routes/map'),
   require('../routes/map/parcel'),
   require('../routes/withdraw'),
