+++ app/routes/v2/search-crn/index.js
@@ -1,5 +1,5 @@
 const joi = require('joi')
-const cache = require('../../../../cache')
+const cache = require('../../../cache')
 
 module.exports = [{
   method: 'GET',
